feat: implement styling
